Move data and domain files to correct path
