+++ data/src/main/java/com/example/bcsd_weather/data/datasource/UltraShortTermLiveDataSource.kt
@@ -1,0 +1,16 @@
+package com.example.bcsd_weather.data.datasource
+
+import kotlinx.coroutines.flow.Flow
+import com.example.bcsd_weather.data.model.UltraShortTermLiveRemote
+
+interface UltraShortTermLiveDataSource {
+    fun getUltraShortTermLive(serviceKey: String,
+                              numOfRows: Int,
+                              pageNo: Int,
+                              dataType: String,
+                              baseDate: String,
+                              baseTime: String,
+                              nx: Int,
+                              ny: Int
+    ): Flow<UltraShortTermLiveRemote>
+}
