+++ data/src/main/java/com/example/bcsd_weather/data/model/UltraShortTermForeCastRemote.kt
@@ -1,0 +1,14 @@
+package com.example.bcsd_weather.data.model
+
+data class UltraShortTermForeCastRemote(val response: UstResponse)
+data class UstResponse(val header: UstHeader, val body: UstBody)
+data class UstHeader(val resultCode: Int, val resultMsg: String)
+data class UstBody(val dataType: String, val items: UstItemList, val totalCount: Int)
+data class UstItemList(val item: List<UltraShortTermFcst>)
+
+data class UltraShortTermFcst(
+    val category: String,   // 자료 구분
+    val fcstDate: String,   // 예측 날짜
+    val fcstTime: String,   // 예측 시간
+    val fcstValue: String,  // 예보 값
+)
