+++ data/src/main/java/com/example/bcsd_weather/data/model/UltraShortTermLiveModel.kt
@@ -1,0 +1,13 @@
+package com.example.bcsd_weather.data.model
+
+import com.google.gson.annotations.SerializedName
+
+// 초단기 실황
+data class UltraShortTermLiveModel (
+    @SerializedName("T1H") var temp: String = "",            // 온도
+    @SerializedName("RN1") var precipitation: String = "",   // 강수량
+    @SerializedName("REH") var humidity: String = "",        // 습도
+    @SerializedName("WSD") var windSpeed: String = "",       // 풍속
+    @SerializedName("fcstTime") var fcstTime: String = ""   // 예보시각
+)
+
