--- domain/src/androidTest/java/model/ShortTermForecast.kt
@@ -1,9 +1,0 @@
-package model
-
-data class ShortTermForecast(
-    val temp: String,
-    val precipitation: String,
-    val humidity: String,
-    val windSpeed: String,
-    val fcstTime: String
-)
