--- domain/src/androidTest/java/model/UltraShortTermForecast.kt
@@ -1,9 +1,0 @@
-package model
-
-data class UltraShortTermForecast(
-    val temp: String,
-    val precipitation: String,
-    val humidity: String,
-    val windSpeed: String,
-    val fcstTime: String
-)
