--- domain/src/androidTest/java/usecase/GetUltraSrtLiveUseCase.kt
@@ -1,26 +1,0 @@
-package usecase
-
-import model.UltraShortTermLive
-import repository.UltraSrtLiveRepository
-
-class GetUltraSrtLiveUseCase(private val ultraSrtLiveRepository: UltraSrtLiveRepository) {
-    suspend fun execute(
-        serviceKey: String,
-        numOfRows: Int,
-        pageNo: Int,
-        dataType: String,
-        baseDate: String,
-        baseTime: String,
-        nx: Int,
-        ny: Int): UltraShortTermLive
-    = ultraSrtLiveRepository.getUltraSrtLive(
-        serviceKey,
-        numOfRows,
-        pageNo,
-        dataType,
-        baseDate,
-        baseTime,
-        nx,
-        ny
-    )
-}
